rearrange components in server/client and parallel routing
