+++ web/src/app/_components/layout/Container.tsx
@@ -1,29 +1,5 @@
-import type { NextPage } from 'next';
-import React from 'react';
-import sizes from '@/config/sizes.json';
+import React from "react";
 
-type Props = {
-  children: React.ReactNode;
-};
-
-const Container: NextPage<Props> = ({ children }) => {
-  return (
-    <div className="container">
-      {children}
-      <style jsx>{`
-        .container {
-          margin: 0 20px;
-        }
-
-        @media (min-width: ${sizes.container}) {
-          .container {
-            width: ${sizes.container};
-            margin: 0 auto;
-          }
-        }
-      `}</style>
-    </div>
-  );
-};
-
-export default Container;
+export default function Container({ children }: { children: React.ReactNode }) {
+  return <div className="container">{children}</div>;
+}
