+++ web/src/app/_components/shared/ProjectItem.tsx
@@ -1,8 +1,5 @@
-import type { NextPage } from 'next';
-import colors from '@/config/colors.json';
-import sizes from '@/config/sizes.json';
-import IconExternalLink from '@/components/svg/IconExternalLink';
-import IconGithub from '@/components/svg/IconGithub';
+import IconExternalLink from "@/components/svg/IconExternalLink";
+import IconGithub from "@/components/svg/IconGithub";
 
 type Item = {
   title: string;
@@ -11,11 +8,7 @@ type Item = {
   link?: string;
 };
 
-type Props = {
-  item: Item;
-};
-
-const ProjectItem: NextPage<Props> = ({ item }) => {
+export default function ProjectItem({ item }: { item: Item }) {
   return (
     <div className="project-item">
       <h4>{item.title}</h4>
@@ -36,28 +29,6 @@ const ProjectItem: NextPage<Props> = ({ item }) => {
           </div>
         )}
       </div>
-      <style jsx>{`
-        p {
-          padding-right: 20px;
-          color: ${colors.light.text_secondary_color};
-        }
-
-        .icons {
-          display: flex;
-          gap: 10px;
-        }
-
-        @media (min-width: ${sizes.container}) {
-        }
-
-        @media (prefers-color-scheme: dark) {
-          p {
-            color: ${colors.dark.text_secondary_color};
-          }
-        }
-      `}</style>
     </div>
   );
-};
-
-export default ProjectItem;
+}
