+++ web/src/app/_components/svg/IconExternalLink.tsx
@@ -1,12 +1,6 @@
-import type { NextPage } from 'next';
-import React from 'react';
-import colors from '@/config/colors.json';
+import React from "react";
 
-type Props = {
-  size: string;
-};
-
-const Svg: NextPage<Props> = ({ size }) => {
+export default function Svg({ size }: { size: string }) {
   return (
     <>
       <svg
@@ -26,27 +20,6 @@ const Svg: NextPage<Props> = ({ size }) => {
         <polyline points="15 3 21 3 21 9"></polyline>
         <line x1="10" y1="14" x2="21" y2="3"></line>
       </svg>
-      <style jsx>{`
-        #icon-external-link {
-          stroke: ${colors.light.text_primary_color};
-        }
-
-        #icon-external-link:hover {
-          stroke: ${colors.light.text_active_color};
-        }
-
-        @media (prefers-color-scheme: dark) {
-          #icon-external-link {
-            stroke: ${colors.dark.text_primary_color};
-          }
-
-          #icon-external-link:hover {
-            stroke: ${colors.dark.text_active_color};
-          }
-        }
-      `}</style>
     </>
   );
-};
-
-export default Svg;
+}
