+++ web/src/app/_components/svg/IconGithub.tsx
@@ -1,12 +1,6 @@
-import type { NextPage } from 'next';
-import React from 'react';
-import colors from '@/config/colors.json';
+import React from "react";
 
-type Props = {
-  size: string;
-};
-
-const Svg: NextPage<Props> = ({ size }) => {
+export default function Svg({ size }: { size: string }) {
   return (
     <>
       <svg
@@ -24,27 +18,6 @@ const Svg: NextPage<Props> = ({ size }) => {
       >
         <path d="M9 19c-5 1.5-5-2.5-7-3m14 6v-3.87a3.37 3.37 0 0 0-.94-2.61c3.14-.35 6.44-1.54 6.44-7A5.44 5.44 0 0 0 20 4.77 5.07 5.07 0 0 0 19.91 1S18.73.65 16 2.48a13.38 13.38 0 0 0-7 0C6.27.65 5.09 1 5.09 1A5.07 5.07 0 0 0 5 4.77a5.44 5.44 0 0 0-1.5 3.78c0 5.42 3.3 6.61 6.44 7A3.37 3.37 0 0 0 9 18.13V22"></path>
       </svg>
-      <style jsx>{`
-        #icon-github {
-          stroke: ${colors.light.text_primary_color};
-        }
-
-        #icon-github:hover {
-          stroke: ${colors.light.text_active_color};
-        }
-
-        @media (prefers-color-scheme: dark) {
-          #icon-github {
-            stroke: ${colors.dark.text_primary_color};
-          }
-
-          #icon-github:hover {
-            stroke: ${colors.dark.text_active_color};
-          }
-        }
-      `}</style>
     </>
   );
-};
-
-export default Svg;
+}
